Make auth

+++ src/Controller/TaskController.php
@@ -3,7 +3,9 @@
 namespace App\Controller;
 
 use App\Entity\Task;
+use App\Entity\User;
 use App\Form\TaskType;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -15,8 +17,11 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks", name="task_list")
      */
-    public function listAction()
+    public function listTasks()
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('homepage');
+        }
         $task = $this->getDoctrine()->getRepository(Task::class)->findBy(['isDone' => 0],['createdAt' => 'DESC']);
         return $this->render('task/list.html.twig', ['tasks' => $task]);
     }
@@ -24,27 +29,28 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/create", name="task_create")
      * @param Request $request
+     * @param EntityManagerInterface $em
      * @return RedirectResponse|Response
      */
-    public function createAction(Request $request)
+    public function createAction(Request $request, EntityManagerInterface $em)
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('homepage');
+        }
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
-
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-
-            $task->isDone(0);
+            $task
+                ->setUser($this->getUser())
+                ->isDone(0);
             $em->persist($task);
             $em->flush();
 
             $this->addFlash('success', 'La tâche a été bien été ajoutée.');
-
             return $this->redirectToRoute('task_list');
         }
-
         return $this->render('task/create.html.twig', ['form' => $form->createView()]);
     }
 
@@ -52,22 +58,23 @@ class TaskController extends AbstractController
      * @Route("/tasks/{id}/edit", name="task_edit")
      * @param Task $task
      * @param Request $request
+     * @param EntityManagerInterface $entityManager
      * @return RedirectResponse|Response
      */
-    public function editAction(Task $task, Request $request)
+    public function editAction(Task $task, Request $request,EntityManagerInterface $entityManager)
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('homepage');
+        }
         $form = $this->createForm(TaskType::class, $task);
-
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $entityManager->flush();
 
             $this->addFlash('success', 'La tâche a bien été modifiée.');
-
             return $this->redirectToRoute('task_list');
         }
-
         return $this->render('task/edit.html.twig', [
             'form' => $form->createView(),
             'task' => $task,
@@ -81,6 +88,9 @@ class TaskController extends AbstractController
      */
     public function toggleTaskAction(Task $task)
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('homepage');
+        }
         $task->toggle(!$task->isDone());
         $this->getDoctrine()->getManager()->flush();
 
@@ -91,7 +101,6 @@ class TaskController extends AbstractController
         $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme non faite.', $task->getTitle()));
         return $this->redirectToRoute('task_list');
         }
-
     }
 
     /**
@@ -99,6 +108,9 @@ class TaskController extends AbstractController
      */
     public function doneTasks()
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('homepage');
+        }
         $task = $this->getDoctrine()->getRepository(Task::class)->findBy(['isDone' => 1],['createdAt' => 'DESC']);
         return $this->render('task/done.html.twig', ['tasks' => $task]);
     }
@@ -110,9 +122,22 @@ class TaskController extends AbstractController
      */
     public function deleteTaskAction(Task $task)
     {
-        // if task->author != $this->getUser();
-        // $this->addFlash('error', 'Vous ne pouvez modifier que les taches que vous avez crées')
-        // redirectToRoute('task_list');
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('homepage');
+        }
+        if ($task->getUser()->getUsername() == 'Anonyme' && $this->getUser()->isAdmin()) {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($task);
+            $em->flush();
+
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+
+            return $this->redirectToRoute('task_list');
+        }
+        if ($task->getUser() != $this->getUser()) {
+            $this->addFlash('error', 'Vous ne pouvez modifier que les taches que vous avez crées');
+            return$this->redirectToRoute('task_list');
+        }
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
@@ -120,5 +145,45 @@ class TaskController extends AbstractController
         $this->addFlash('success', 'La tâche a bien été supprimée.');
 
         return $this->redirectToRoute('task_list');
+    }
+
+    // *** PART OF TASKS CONTROLLER FOR ROLE_ADMIN ONLY *** //
+
+    /**
+     * @Route("/admin/anotasks", name="task_anonymous")
+     * @return Response
+     */
+    public function anonymousTaskList()
+    {
+        $user = $this->getUser();
+        if ($user) {
+            if ($user->isAdmin()) {
+                $tasks = $this->getDoctrine()->getRepository(Task::class)->findBy(['user' => null]);
+                return $this->render('task/anonymoustask.html.twig', ['tasks'=> $tasks]);
+            }
+        }
+        return $this->redirectToRoute('homepage');
+    }
+
+    /**
+     * @Route("/admin/taskto/{id}", name="task_change")
+     * @param Task $task
+     * @param EntityManagerInterface $entityManager
+     * @return RedirectResponse
+     */
+    public function taskToAnonymous(Task $task, EntityManagerInterface $entityManager)
+    {
+        $user = $this->getUser();
+        if ($user) {
+            if ($user->isAdmin) {
+                $anonymous = $this->getDoctrine()->getRepository(User::class)->findOneBy(['username' => 'Anonyme']);
+                $task->setUser($anonymous);
+                $entityManager->persist($task);
+                $entityManager->flush();
+                $this->addFlash('success', 'L\'auteur de la tache a bien été modifié');
+                return $this->redirectToRoute('task_anonymous');
+            }
+        }
+        return $this->redirectToRoute('homepage');
     }
 }
